Made player stats sync w/ playerSave and added super speed during debug

+++ topdowngame/PlayerSave.h
@@ -13,7 +13,7 @@ public:
 
 	int playerXP;
 
-	int baseStats[4]; //HP, Attack, Defense, Speed
+	double baseStats[4]; //HP, Attack, Defense, Speed
 
 	PlayerSave();
